% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db_connect.R
\name{access_db_connect}
\alias{access_db_connect}
\title{Establish a connection to an MS Access database}
\usage{
access_db_connect(file_path = NULL)
}
\arguments{
\item{file_path}{full path to the MS Access file
The default will be the Access template database}
}
\value{
A connection object to the MS ACCESS database.
}
\description{
This is a wrapper for the `DBI::dbConnect`, see it's documentation for more details.
}
