% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{add_hard_coded_values}
\alias{add_hard_coded_values}
\title{Add hard-coded or computed values}
\usage{
add_hard_coded_values(df)
}
\arguments{
\item{df}{Input dataframe}
}
\value{
Dataframe with additional columns
}
\description{
Add hard-coded or computed values
}
