% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andes_db_connect.R
\name{andes_db_connect}
\alias{andes_db_connect}
\title{Establish a connection to the andes database}
\usage{
andes_db_connect(url_bd, port_bd, nom_usager, mot_de_passe, nom_bd = "andesdb")
}
\arguments{
\item{url_bd}{URL of the ANDES database server.}

\item{port_bd}{Port number for the ANDES database connection.}

\item{nom_usager}{Username for the ANDES database.}

\item{mot_de_passe}{Password for the ANDES database.}

\item{nom_bd}{Name of the ANDES database. Default is "andesdb".}
}
\value{
A connection object to the ANDES database.
}
\description{
This is a wrapper for the `DBI::dbConnect`, see it's documentation for more details.
}
