% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{andes_str_to_oracle_datetime}
\alias{andes_str_to_oracle_datetime}
\title{Convert andes datetime string form DB to Oracle datetime format}
\usage{
andes_str_to_oracle_datetime(datetime_str)
}
\arguments{
\item{datetime_str}{The string representing the datetime from ANDES DB}
}
\description{
Convert andes datetime string form DB to Oracle datetime format
}
