% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{check_numeric_columns}
\alias{check_numeric_columns}
\title{Make sure the columns listed in col_names have numeric (or NA) values}
\usage{
check_numeric_columns(df, col_names = NULL, coerce = FALSE)
}
\arguments{
\item{df}{the dataframe to verify}

\item{col_names}{A list of column names. This will veridy if columns not in the list is present}

\item{coerce}{A boolean (false by default) to see if the dataframe can be coerced into compliance}
}
\value{
A boolean representing if the dataframe is compliant.
}
\description{
Make sure the columns listed in col_names have numeric (or NA) values
}
