% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{check_other_columns}
\alias{check_other_columns}
\title{Make sure no other columns than the ones listed in col_names are present in the dataframe}
\usage{
check_other_columns(df, col_names = NULL, coerce = FALSE)
}
\arguments{
\item{df}{the dataframe to verify}

\item{col_names}{A list of column names. This will veridy if columns not in the list is present}

\item{coerce}{A boolean (false by default) to see if the dataframe can be coerced into compliance}
}
\value{
A boolean representing if the dataframe is compliant.
}
\description{
Make sure no other columns than the ones listed in col_names are present in the dataframe
}
