% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{cleanup_text}
\alias{cleanup_text}
\title{Cleanup text form ANDES to play nice in Oracle}
\usage{
cleanup_text(df, col_name = NULL, max_chars = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{col_name}{String the name of the column to target in the dataframe}

\item{max_chars}{Numeric, maximum characters allowed in Oracle}
}
\description{
Cleanup text form ANDES to play nice in Oracle
}
