% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{cols_contains_na}
\alias{cols_contains_na}
\title{Checks if all dataframe cols named in the col_names contain NA
This is useful to validate if a dataframe can be written to a DB table (where some columns values cannot be null)}
\usage{
cols_contains_na(df, col_names = NULL)
}
\arguments{
\item{df:}{the datafram to modify}

\item{col_names:}{a list of column names which will be converted to numeric}
}
\description{
Checks if all dataframe cols named in the col_names contain NA
This is useful to validate if a dataframe can be written to a DB table (where some columns values cannot be null)
}
