% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{cols_to_numeric}
\alias{cols_to_numeric}
\title{Convert all dataframe cols named in the col_names to a numeric value}
\usage{
cols_to_numeric(df, col_names = NULL)
}
\arguments{
\item{df}{the dataframe to modify}

\item{col_names}{a list of column names which will be converted to numeric}
}
\description{
Convert all dataframe cols named in the col_names to a numeric value
}
