% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db_connect.R
\name{create_new_access_db}
\alias{create_new_access_db}
\title{Creates a new MS Access database file}
\usage{
create_new_access_db(fname = "new_access_db.mdb")
}
\arguments{
\item{fname}{Filename for new access DB}
}
\description{
Creates a new MS Access database file
}
