% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{determine_time_type}
\alias{determine_time_type}
\title{Determine time type (standard or daylight saving)}
\usage{
determine_time_type(datetime)
}
\arguments{
\item{datetime}{Input datetime}
}
\value{
Time type code
}
\description{
Determine time type (standard or daylight saving)
}
