% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{determine_timezone}
\alias{determine_timezone}
\title{Determine timezone}
\usage{
determine_timezone(datetime)
}
\arguments{
\item{datetime}{Input datetime}
}
\value{
Timezone code
}
\description{
Determine timezone
}
