% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_capture_mollusque.R
\name{format_cod_esp_gen}
\alias{format_cod_esp_gen}
\title{format the cod_esp_gen column}
\usage{
format_cod_esp_gen(capt)
}
\arguments{
\item{capt}{Dataframe de capture}
}
\description{
format the cod_esp_gen column
}
