% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{format_cod_secteur_releve}
\alias{format_cod_secteur_releve}
\title{Format COD_SECTEUR_RELEVE}
\usage{
format_cod_secteur_releve(trait, desc_secteur_releve_f)
}
\arguments{
\item{trait}{Dataframe du trait}

\item{desc_secteur_releve_f}{String a associer au trait}
}
\description{
Format COD_SECTEUR_RELEVE
}
