% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{format_cod_strate}
\alias{format_cod_strate}
\title{Format COD_STRAT}
\usage{
format_cod_strate(trait, desc_serie_hist_f)
}
\arguments{
\item{trait}{Dataframe}

\item{desc_serie_hist_f}{String}
}
\description{
Format COD_STRAT
}
