% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_freq_long_mollusque.R
\name{format_cod_tech_mesure_long}
\alias{format_cod_tech_mesure_long}
\title{Format COD_TECH_MESURE_LONG}
\usage{
format_cod_tech_mesure_long(freq)
}
\arguments{
\item{freq}{Dataframe to reformat}
}
\description{
Format COD_TECH_MESURE_LONG
}
