% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_freq_long_mollusque.R
\name{format_cod_typ_etat}
\alias{format_cod_typ_etat}
\title{Format COD_TYP_ETAT}
\usage{
format_cod_typ_etat(freq)
}
\arguments{
\item{freq}{Dataframe to reformat}
}
\description{
Format COD_TYP_ETAT
}
