% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{format_cod_typ_heure}
\alias{format_cod_typ_heure}
\title{Format COD_TYP_HEURE}
\usage{
format_cod_typ_heure(trait)
}
\arguments{
\item{trait}{Dataframe}
}
\description{
Format COD_TYP_HEURE
}
