% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_freq_long_mollusque.R
\name{format_cod_typ_long}
\alias{format_cod_typ_long}
\title{Format COD_TYP_LONG}
\usage{
format_cod_typ_long(freq)
}
\arguments{
\item{freq}{Dataframe to reformat}
}
\description{
Format COD_TYP_LONG
}
