% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{format_coordinates}
\alias{format_coordinates}
\title{Format coordinates for TRAIT_MOLLUSQUE}
\usage{
format_coordinates(trait)
}
\arguments{
\item{trait}{Dataframe}
}
\value{
Formatted dataframe
}
\description{
Converts coordinates to Oracle-specific format
}
