% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_capture_mollusque.R
\name{format_epibiont}
\alias{format_epibiont}
\title{Formated COD_COUVERTURE_EPIBIONT column to the dataframe
This categorizes AND includes the of the following definition:
ave_coverage to cod_couverture :
0 -> Aucune balane
1 -> 1/3 et moins surface colonisée
2 -> 1/3 à 2/3 surface colonisée
3 -> 2/3 et plus surface colonisée}
\usage{
format_epibiont(andes_db_connection, code_filter)
}
\arguments{
\item{andes_db_connection}{A connection object to the ANDES database.}

\item{code_filter}{a list of species code to filter on, or NULL for no filtering}
}
\value{
The input dataframe with columns for categorical codes
}
\description{
#' ave_with_barnacles to cod_abondance
0 -> Aucun des pétoncles ne porte de balane
1 -> 1% à 20%% des pétoncles portent des balanes
2 -> 21% à 40%% des pétoncles portent des balanes
3 -> 41% à 60%% des pétonlces portent des balanes
4 -> 61% à 80%% despétoncles portent des balanes
5 -> 81% à 100%% des pétonlces portent des balanes
}
