% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_freq_long_mollusque.R
\name{format_no_mollusque}
\alias{format_no_mollusque}
\title{Format NO_MOLLUSQUE}
\usage{
format_no_mollusque(freq)
}
\arguments{
\item{freq}{Dataframe to reformat}
}
\description{
Format NO_MOLLUSQUE
}
