% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{format_zone}
\alias{format_zone}
\title{Format ZONE}
\usage{
format_zone(trait, desc_serie_hist_f)
}
\arguments{
\item{trait}{Dataframe}

\item{desc_serie_hist_f}{String}
}
\description{
Format ZONE
}
