% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{format_zone_secteur}
\alias{format_zone_secteur}
\title{Format zone and sector information}
\usage{
format_zone_secteur(df)
}
\arguments{
\item{df}{Input dataframe}
}
\value{
Formatted dataframe
}
\description{
Format zone and sector information
}
