% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_mollusque.R
\name{get_capture_mollusque}
\alias{get_capture_mollusque}
\title{Gets capture_mollusque (formatted results)}
\usage{
get_capture_mollusque(
  andes_db_connection,
  engin = NULL,
  code_filter = NULL,
  basket_class_filter = NULL
)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}

\item{code_filter}{Optional list of species to filter}

\item{basket_class_filter}{Optional list of basket class to filter}
}
\value{
A dataframe containing capture_mollusque table data.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the CAPTURE_MOLLUSC table.
The current ANDES active mission will determine for which data are returned.
}
\details{
Structurally, it would make sense to send a Trait instance to for every capture_mollusque
but here we can get away with proj (due to how ANDES is structured)
}
\seealso{
[get_capture_mollusque_db()] for the db results
}
