% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_mollusque.R
\name{get_capture_mollusque_db}
\alias{get_capture_mollusque_db}
\title{Gets capture_mollusc_db (raw database results)}
\usage{
get_capture_mollusque_db(
  andes_db_connection,
  code_filter = NULL,
  basket_class_filter = NULL
)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}

\item{code_filter}{Optional list of species to filter}

\item{basket_class_filter}{Optional list of basket class to filter}
}
\value{
A dataframe containing fishing set data.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the CAPTURE_MOLLUSC table.
The current ANDES active mission will determine for which data are returned.
}
\details{
This function is intended for internal use and returns raw results from the database.
It is not meant for direct use in analysis or reporting. Users should use `get_capture_mollusc`
}
\seealso{
[get_capture_mollusque()] for the formatted results
}
