% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engin_mollusque.R
\name{get_engin_mollusque_db}
\alias{get_engin_mollusque_db}
\title{Gets engin_mollusc_db (raw database results)}
\usage{
get_engin_mollusque_db(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
A dataframe containing fishing set data.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the ENGIN_MOLLUSC table.
The current ANDES active mission will determine for which data are returned.
}
\details{
This function is intended for internal use and returns raw results from the database.
It is not meant for direct use in analysis or reporting. Users should use `get_engin_mollusc`
}
\seealso{
[get_projet_mollusque()] for the formatted results
}
