% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_long_mollusque.R
\name{get_freq_long_mollusque}
\alias{get_freq_long_mollusque}
\title{Gets freq_long_mollusque (formatted results)}
\usage{
get_freq_long_mollusque(andes_db_connection, capt = NULL)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}

\item{capt}{A formatted capture_mollusque dataframe}
}
\value{
A dataframe containing formatted fishing set data.
}
\description{
This function executes a SQL query to retrieve the needed andes data
to construct the TRAIT_MOLLUSQUE table and formats the results.
}
