% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_long_mollusque.R
\name{get_freq_long_mollusque_db}
\alias{get_freq_long_mollusque_db}
\title{Gets freq_long_mollusque_db (raw database results)}
\usage{
get_freq_long_mollusque_db(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
A dataframe containing fishing set data.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the FREQ_LONG_MOLLUSQUE table.
The current ANDES active mission will determine for which data are returned.
}
\details{
This function is intended for internal use and returns raw results from the database.
It is not meant for direct use in analysis or reporting. Users should use `get_freq_long_mollusque`
}
\seealso{
[get_freq_long_mollusque()] for the formatted results
}
