% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometrie_petoncle.R
\name{get_legal_collection_names}
\alias{get_legal_collection_names}
\title{Get a list of legal collection names as a filter for get_biometrie_petoncle()}
\usage{
get_legal_collection_names()
}
\value{
A dataframe containing get_biometrie_petoncle table data.
}
\description{
These must match the shared_models_observationgroup.nom as the ANDES collection.
The secteur (16E, 16F, centre, ouest) are taken from the last part of the string.
}
