% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_mollusc.R
\name{get_projet_mollusc}
\alias{get_projet_mollusc}
\title{Gets fishing projet_mollusc (raw database results)}
\usage{
get_projet_mollusc(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database. Please}
}
\value{
A dataframe containing fishing set data.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the PROJET_MOLLUSC table.
The current ANDES active mission will determine for which data are returned.
}
\seealso{
[get_projet_mollusc_db()] for the raw database results
}
