% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{get_strate}
\alias{get_strate}
\title{Get the strate from reference}
\usage{
get_strate(nom_station, desc_serie_hist_f)
}
\arguments{
\item{nom_station}{String}

\item{desc_serie_hist_f}{String}
}
\description{
Get the strate from reference
}
