% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_mollusque.R
\name{get_trait_mollusque}
\alias{get_trait_mollusque}
\title{Gets trait_mollusque (formatted results)}
\usage{
get_trait_mollusque(andes_db_connection, proj = NULL)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}

\item{proj}{A formatted projet_mollusque dataframe}
}
\value{
A dataframe containing formatted fishing set data.
}
\description{
This function executes a SQL query to retrieve the needed andes data 
to construct the TRAIT_MOLLUSQUE table and formats the results.
}
