% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{is_andes_time_str_dst}
\alias{is_andes_time_str_dst}
\title{Verify is the ANDES dattime string is in daylight savings time}
\usage{
is_andes_time_str_dst(datetime_str)
}
\arguments{
\item{datetime_str}{The string representing the datetime from ANDES DB}
}
\value{
A boolean representing if the time is in DST
}
\description{
Verify is the ANDES dattime string is in daylight savings time
}
