% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{left_join}
\alias{left_join}
\title{a merge that preserves row and column order}
\usage{
left_join(x, y, ...)
}
\arguments{
\item{x, }{the "left" dataframe (all.x=TRUE)}

\item{y, }{the "right" dataframe}

\item{by, }{by The name (or names) of the column to merge on, must be in both x and y.}
}
\description{
shamelessly stolen from https://stackoverflow.com/questions/17878048/merge-two-data-frames-while-keeping-the-original-row-order
}
