% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_projet_mollusque.R
\name{lookup_cod_serie_hist}
\alias{lookup_cod_serie_hist}
\title{Lookup entry in Oracle}
\usage{
lookup_cod_serie_hist(desc_serie_hist_f)
}
\arguments{
\item{desc_serie_hist_f}{String value}
}
\description{
Lookup entry in Oracle
}
