% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{lookup_station}
\alias{lookup_station}
\title{This requires opening station reference data to determine the zone.}
\usage{
lookup_station(station_name = NULL, zone = NULL, species = NULL)
}
\description{
This requires opening station reference data to determine the zone.
}
