% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{sanitize_sql_value}
\alias{sanitize_sql_value}
\title{Sanitize value to SQL statement}
\usage{
sanitize_sql_value(value)
}
\arguments{
\item{value}{The value to sanitize}
}
\description{
It will wrap string with an extra set of single quotes.
It will escape every single quote by doubling it up
This usualy does nothing to the value itself except inject the NULL string for NA/null and empty strings
}
