% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{validate_set_result}
\alias{validate_set_result}
\title{Validate set result consistency}
\usage{
validate_set_result(trait)
}
\arguments{
\item{trait}{Input dataframe}
}
\value{
Formatted dataframe
}
\description{
ANDES has a mecanism to tag valid sets, but we also inserted a
legacy set observation for a similar reason. This functions makes
sure both point in the same direction and prints an error code if
there are contradictions.
}
