% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_mollusque.R
\name{write_capture_mollusque}
\alias{write_capture_mollusque}
\title{Write to Access file}
\usage{
write_capture_mollusque(df, access_db_write_connection = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{access_db_write_connection}{the connection to Access file}
}
\description{
Write to Access file
}
