% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_long_mollusque.R
\name{write_freq_long_mollusque}
\alias{write_freq_long_mollusque}
\title{Write dataframe to Access file}
\usage{
write_freq_long_mollusque(df, access_db_write_connection = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{access_db_write_connection}{connection to Access file}
}
\description{
Write dataframe to Access file
}
