% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_mollusque.R
\name{write_projet_mollusque}
\alias{write_projet_mollusque}
\title{Write to Access file}
\usage{
write_projet_mollusque(proj, access_db_write_connection = NULL)
}
\arguments{
\item{proj}{Dataframe}

\item{access_db_write_connection}{the connection to Access file}
}
\description{
Write to Access file
}
