% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_mollusque.R
\name{write_trait_mollusque}
\alias{write_trait_mollusque}
\title{Write dataframe to Access file}
\usage{
write_trait_mollusque(trait, access_db_write_connection = NULL)
}
\arguments{
\item{trait}{Dataframe}

\item{access_db_write_connection}{the connection to Access file}
}
\description{
Write dataframe to Access file
}
