% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{add_hard_coded_value}
\alias{add_hard_coded_value}
\title{Add a hard-coded column with a specific value to the dataframe}
\usage{
add_hard_coded_value(df, col_name = NULL, value = NULL)
}
\arguments{
\item{df}{The original dataframe to modify with a new column}

\item{col_name}{The new column name}

\item{value}{The value to add to every row in this column. 
To add null values in the column use NA and not NULL.}
}
\value{
The original dataframe with the new column added
}
\description{
Add a hard-coded column with a specific value to the dataframe
}
