% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{check_columns_present}
\alias{check_columns_present}
\title{Make sure the columns listed in col_names are present in the dataframe}
\usage{
check_columns_present(df, col_names = NULL, coerce = FALSE)
}
\arguments{
\item{df:}{the dataframe to verify}

\item{col_names:}{A list of column names. This will verify if the names in the list are present.}

\item{coerce:}{A boolean (false by default) to see if the dataframe can be coerced into compliance}
}
\value{
A boolean representing if the dataframe is compliant.
}
\description{
Make sure the columns listed in col_names are present in the dataframe
}
