% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_capture_mollusque.R
\name{format_cod_typ_mesure}
\alias{format_cod_typ_mesure}
\title{For now, we will just assume all data is quantitative
since we are limiting ourselves to commercial (scallops and whelk).
One day, this function should be generalized to also lookup the data and determine it.}
\usage{
format_cod_typ_mesure(capt)
}
\description{
For now, we will just assume all data is quantitative
since we are limiting ourselves to commercial (scallops and whelk).
One day, this function should be generalized to also lookup the data and determine it.
}
