% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{format_date_hre_trait}
\alias{format_date_hre_trait}
\title{Format dates for TRAIT_MOLLUSQUE}
\usage{
format_date_hre_trait(trait)
}
\arguments{
\item{df}{Input dataframe}
}
\value{
Formatted dataframe
}
\description{
Converts and time fields to appropriate format uses andes_str_to_oracle_datetime
}
