% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{format_date_trait}
\alias{format_date_trait}
\title{Format dates for TRAIT_MOLLUSQUE}
\usage{
format_date_trait(trait_df)
}
\arguments{
\item{df}{Input dataframe}
}
\value{
Formatted dataframe
}
\description{
Converts date and time fields to appropriate format
}
