% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{generate_sql_insert_statement}
\alias{generate_sql_insert_statement}
\title{generate a SQL inster statement for the single dataframe row as a new row into table_name
The dataframe must have named columns that correspond to the columns of the table
The values must have the correct data types t(here will be some SQL value sanitizing)
The statement will look like:" INSERT INTO {table_name} {col_names_str} VALUES {col_values_str}"
Where {col_names_str} is list of column names with parentheses: "(NO_RELEVE COD_NBPC ANNEE COD_TYP_STRATIF DATE_DEB_PROJET...)"
and {col_values_str} is list of column values with parentheses: "(36 4 2025 7 '2025-05-03'...)"}
\usage{
generate_sql_insert_statement(df_row, table_name)
}
\description{
generate a SQL inster statement for the single dataframe row as a new row into table_name
The dataframe must have named columns that correspond to the columns of the table
The values must have the correct data types t(here will be some SQL value sanitizing)
The statement will look like:" INSERT INTO {table_name} {col_names_str} VALUES {col_values_str}"
Where {col_names_str} is list of column names with parentheses: "(NO_RELEVE COD_NBPC ANNEE COD_TYP_STRATIF DATE_DEB_PROJET...)"
and {col_values_str} is list of column values with parentheses: "(36 4 2025 7 '2025-05-03'...)"
}
