% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometrie_petoncle.R
\name{get_biometrie_petoncle}
\alias{get_biometrie_petoncle}
\title{Gets get_biometrie_petoncle (formatted results)}
\usage{
get_biometrie_petoncle(andes_db_connection, collection_name = NULL)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
A dataframe containing get_biometrie_petoncle table data.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the biometry table.
The current ANDES active mission will determine for which data are returned.
}
\seealso{
[get_biometrie_petoncle_db(), get_legal_collection_names()] for the db results
}
