% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometrie_petoncle.R
\name{get_biometrie_petoncle_db}
\alias{get_biometrie_petoncle_db}
\title{Gets capture_mollusc_db (raw database results)}
\usage{
get_biometrie_petoncle_db(andes_db_connection, collection_name = NULL)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}

\item{collection_name}{Optional. A string with the name of the colelction, (e.g., "Conserver pour biométrie 16E").}
}
\value{
A dataframe containing fishing set data.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the biometry table.
The current ANDES active mission will determine for which data are returned.
}
\details{
This function is intended for internal use and returns raw results from the database.
It is not meant for direct use in analysis or reporting. Users should use `get_biometrie_petoncle`
}
\seealso{
[get_biometrie_petoncle()] for the formatted results
}
