% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engin_mollusque.R
\name{get_engin_mollusque}
\alias{get_engin_mollusque}
\title{Gets engin_mollusc (formatted results)}
\usage{
get_engin_mollusque(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
A dataframe containing fishing set data.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the ENGIN_MOLLUSC table.
The current ANDES active mission will determine for which data are returned.
}
\seealso{
[get_projet_mollusque_db()] for the db results
}
