% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_capture_mollusque.R
\name{get_epibiont}
\alias{get_epibiont}
\title{This fetches the specimen-level coverage data and
coverts it to an average set-level metric in accordance to the legacy Oracle database.}
\usage{
get_epibiont(andes_db_connection, code_filter)
}
\description{
This fetches the specimen-level coverage data and
coverts it to an average set-level metric in accordance to the legacy Oracle database.
}
