% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db_connect.R
\name{get_ref_choices}
\alias{get_ref_choices}
\title{Builds a list of legal choices (descriptions) for get ref key}
\usage{
get_ref_choices(table = "tablename", col = "columnname", optional_query = "")
}
\description{
Builds a list of legal choices (descriptions) for get ref key
}
