% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db_connect.R
\name{get_ref_key}
\alias{get_ref_key}
\title{Get the reference key corresponding to a value (usually from the Oracle / MSAccess reference database)}
\usage{
get_ref_key(
  table = "tablename",
  pkey_col = "columnofprimarykey",
  col = "columnname",
  val = "entryvalue",
  optional_query = ""
)
}
\arguments{
\item{table:}{The name of the Oracle table, defaults to "tablename"}

\item{pkey_col:}{The column name that holds the key, defaults to "columnofprimarykey"}

\item{col:}{The column that holds the value to match, defaults to "columnname"}

\item{val:}{The value to match, defaults to "entryvalue"}

\item{optional_query:}{additional string to append to the query}
}
\value{
: The value found in the pkey column for the entry with the value
}
\description{
Get the reference key corresponding to a value (usually from the Oracle / MSAccess reference database)
}
