% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{parse_andes_datetime}
\alias{parse_andes_datetime}
\title{takes a standard ANDES UTC time string and converts it to a POSIXct object}
\usage{
parse_andes_datetime(andes_time_str)
}
\description{
takes a standard ANDES UTC time string and converts it to a POSIXct object
}
