% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{strip_alphabetic}
\alias{strip_alphabetic}
\title{This function removes alphabetic characters from a string}
\usage{
strip_alphabetic(my_string)
}
\description{
This function removes alphabetic characters from a string
}
