% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{to_oracle_coord}
\alias{to_oracle_coord}
\title{Convert coordinate to Oracle format}
\usage{
to_oracle_coord(coord)
}
\arguments{
\item{coord}{Input coordinate}
}
\value{
Formatted coordinate
}
\description{
For example, the latitude of 47.155927
       is decomposed into:
       whole_degrees = 47
       decimal_minues = 35562
       and yields: 4709.35562
}
