% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_trait_mollusque.R
\name{to_oracle_coord}
\alias{to_oracle_coord}
\title{Convert coordinate to Oracle format}
\usage{
to_oracle_coord(coord)
}
\arguments{
\item{coord}{Input coordinate}
}
\value{
Formatted coordinate
}
\description{
Convert coordinate to Oracle format
}
