% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_long_mollusque.R
\name{validate_freq_long_mollusque}
\alias{validate_freq_long_mollusque}
\title{Perform database validation checks on the dataframe}
\usage{
validate_freq_long_mollusque(df)
}
\arguments{
\item{df}{The dataframe to validate}
}
\value{
Boolean representing if all the validation tests have passed
}
\description{
Perform database validation checks on the dataframe
}
\details{
This compares the dataframe columns and values to the requirements of the database
}
